requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
openai==1.1.1
colorama==0.4.6
tqdm==4.66.1
Pillow==10.0.0
reportlab==3.6.13
lxml==4.9.3
urllib3==2.0.7
flask==2.3.3
streamlit==1.27.2
wordpress-api==1.2.9
python-whois==0.8.0
validators==0.22.0
spacy==3.7.2
click==8.1.7
pyyaml==6.0.1
pandas==2.1.1
matplotlib==3.8.0
selenium==4.15.2
webdriver-manager==4.0.1
googlesearch-python==1.2.3
fake-useragent==1.3.0
fpdf==1.7.2
psutil==5.9.6
# Nuevas dependencias para el sistema de monitorización
apscheduler==3.10.1
weasyprint==59.0
jinja2==3.1.2
email-validator==2.0.0.post2
plotly==5.18.0
pytz==2023.3
flask-mail==0.9.1
